Se agregan los endpoints para borrar una pelicula de favoritos y traer el listado de peliculas favoritas de un usuario

+++ MoviesApp_BackEnd/Models/UsersFavoriteMovies.cs
@@ -6,7 +6,7 @@ namespace MoviesApp_BackEnd.Models
     {
         [Key]
         public long Id { get; set; }
-        public int UserId { get; set; }
-        public int MovieId { get; set; }
+        public long UserId { get; set; }
+        public long MovieId { get; set; }
     }
 }
